external info component added, result entry must be updated

+++ server/controllers/studentController.js
@@ -1,8 +1,28 @@
 const Student = require("../models/student");
 const ExternalStudentPriorInfo = require('../models/external-student-info');
+const StudentGrade = require("../models/student-grade");
 //const mongoose = require('mongoose');
 
 const studentController = {
+
+    getAllStudents: async (req, res) => {
+        try {
+            const students = await Student.find();
+            res.status(200).json(students);
+        } catch (error) {
+            res.status(500).json({ message: "Error fetching students", error });
+        }
+    },
+
+    getNewStudents: async (req, res) => {
+        try {
+            const students = await Student.find({ has_perior_school_info: false, registered: false });
+            res.status(200).json(students);
+        } catch (error) {
+            res.status(500).json({ message: "Error fetching students", error });
+        }
+    },
+
     createStudent: async (req, res) => {
         //const session = await mongoose.startSession();
         try {
@@ -31,26 +51,21 @@ const studentController = {
         }
     },
 
-
-    getAllStudents: async (req, res) => {
+    updateStudent: async (req, res) => {
         try {
-            const students = await Student.find();
-            res.status(200).json(students);
+            const { id } = req.params;
+            const { first_name, middle_name, last_name, sex, birth_date } = req.body;
+            const updatedStudent = await Student.findByIdAndUpdate(id, { first_name, middle_name, last_name, sex, birth_date }, { new: true });
+            if (!updatedStudent) {
+                return res.status(404).json({ message: "Student not found" });
+            }
+            res.status(200).json(updatedStudent);
         } catch (error) {
-            res.status(500).json({ message: "Error fetching students", error });
+            res.status(500).json({ message: error.message });
         }
     },
 
-    getNewStudents: async (req, res) => {
-        try {
-            const students = await Student.find({ has_perior_school_info: false, registered: false });
-            res.status(200).json(students);
-        } catch (error) {
-            res.status(500).json({ message: "Error fetching students", error });
-        }
-    },
-
-
+    /*
     updateStudent: async (req, res) => {
         try {
             const { id } = req.params;
@@ -84,12 +99,17 @@ const studentController = {
             res.status(500).json({ message: error.message });
         }
     },
-
+    */
     deleteStudent: async (req, res) => {
         try {
             const { id } = req.params;
+            const studExists = await StudentGrade.exists({ student: id });
+            if (studExists) {
+                return res.status(400).json({
+                    message: "Cannot delete the student. It is associated with one or more class.",
+                });
+            }
             const student = await Student.findByIdAndDelete(id);
-
             if (!student) {
                 return res.status(404).json({ message: "Student not found" });
             }
